fix overflow

+++ projects/rccl/src/include/algorithms/CollCommon.h
@@ -116,8 +116,8 @@ static inline __device__ void reduceScatter(T* const* __restrict__ ipcbuffs, T* 
   for (size_t idx = idxStart; idx < idxEnd; idx += idxStride) {
     // pattern = 2 performs reduce (one-shot)
     // pattern = 1 performs reduce-scatter (two-shot)
-    size_t srcIdx = (pattern == 2) ? idx : (idx + selfRank * idxEnd);
-    size_t destIdx = (pattern == 1) ? (idx + selfRank * idxEnd) : idx;
+    size_t srcIdx = (pattern == 2) ? idx : (idx + static_cast<size_t>(selfRank) * idxEnd);
+    size_t destIdx = (pattern == 1) ? (idx + static_cast<size_t>(selfRank) * idxEnd) : idx;
 
     uint4 sum{0, 0, 0, 0};
     if constexpr (hasAcc) {
@@ -150,7 +150,7 @@ static inline __device__ void allGather(T* const* __restrict__ ipcbuffs, T* __re
 #pragma unroll kUnroll
     for (int r = 0; r < nRanks; ++r) {
       int srcRank = (selfRank + r) % nRanks;
-      size_t destIdx = idx + srcRank * idxEnd;
+      size_t destIdx = idx + static_cast<size_t>(srcRank) * idxEnd;
       size_t srcIdx;
       if (enable_offset) {
         srcIdx = destIdx;
